chore :: interceptor도입으로 인한 header annotation 제거
