--- app/src/main/java/team/devlib/android/data/api/UserApi.kt
@@ -1,28 +1,0 @@
-package team.devlib.android.data.api
-
-import retrofit2.http.Body
-import retrofit2.http.GET
-import retrofit2.http.Header
-import retrofit2.http.POST
-import team.devlib.android.data.model.user.request.SignInRequest
-import team.devlib.android.data.model.user.request.SignUpRequest
-import team.devlib.android.data.model.user.response.TokenResponse
-import team.devlib.android.data.model.user.response.UserInformationResponse
-import team.devlib.android.data.util.RequestUrl
-
-interface UserApi {
-    @POST(RequestUrl.User.auth)
-    suspend fun signIn(
-        @Body signInRequest: SignInRequest,
-    ): TokenResponse
-
-    @POST(RequestUrl.User.user)
-    suspend fun signUp(
-        @Body signUpRequest: SignUpRequest,
-    ): TokenResponse
-
-    @GET(RequestUrl.User.user)
-    suspend fun fetchUserInformation(
-        @Header("Authorization") token: String,
-    ): UserInformationResponse
-}
