+++ app/src/main/java/team/devlib/android/data/remote/api/BookApi.kt
@@ -1,28 +1,25 @@
-package team.devlib.android.data.api
+package team.devlib.android.data.remote.api
 
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
-import team.devlib.android.data.model.book.FetchBookRankingResponse
-import team.devlib.android.data.model.book.FetchMyBookmarksResponse
+import team.devlib.android.data.remote.model.book.FetchBookRankingResponse
+import team.devlib.android.data.remote.model.book.FetchMyBookmarksResponse
 import team.devlib.android.data.util.RequestUrl
 import team.devlib.android.feature.home.Type
 
 interface BookApi {
     @GET(RequestUrl.Book.mark)
     suspend fun fetchMyBookmarks(
-        @Header("Authorization") token: String,
     ): FetchMyBookmarksResponse
 
     @GET(RequestUrl.Book.rank)
     suspend fun fetchBookRanking(
-        @Header("Authorization") token: String,
-        @Query("type") type: Type,
+        @Query("type") type: Type
     ): FetchBookRankingResponse
 
     @GET(RequestUrl.Book.book)
     suspend fun searchBook(
-        @Header("Authorization") token: String,
-        @Query("name") name: String,
+        @Query("name") name: String
     ): FetchBookRankingResponse
 }
